+++ src/core/type/type_info.cc
@@ -57,22 +57,22 @@ ListType Type::as_list_type() const
   return ListType{impl()};
 }
 
-void Type::record_reduction_operator(int32_t op_kind, int64_t global_op_id) const
+void Type::record_reduction_operator(int32_t op_kind, int32_t global_op_id) const
 {
   impl()->record_reduction_operator(op_kind, global_op_id);
 }
 
-void Type::record_reduction_operator(ReductionOpKind op_kind, int64_t global_op_id) const
+void Type::record_reduction_operator(ReductionOpKind op_kind, int32_t global_op_id) const
 {
   impl()->record_reduction_operator(static_cast<int32_t>(op_kind), global_op_id);
 }
 
-int64_t Type::find_reduction_operator(int32_t op_kind) const
+int32_t Type::find_reduction_operator(int32_t op_kind) const
 {
   return impl()->find_reduction_operator(op_kind);
 }
 
-int64_t Type::find_reduction_operator(ReductionOpKind op_kind) const
+int32_t Type::find_reduction_operator(ReductionOpKind op_kind) const
 {
   return impl()->find_reduction_operator(static_cast<int32_t>(op_kind));
 }
