+++ src/core/type/type_info.h
@@ -159,7 +159,7 @@ class Type {
    * @param op_kind Reduction operator kind
    * @param global_op_id Global reduction operator ID
    */
-  void record_reduction_operator(int32_t op_kind, int64_t global_op_id) const;
+  void record_reduction_operator(int32_t op_kind, int32_t global_op_id) const;
   /**
    * @brief Records a reduction operator.
    *
@@ -169,7 +169,7 @@ class Type {
    * @param op_kind Reduction operator kind
    * @param global_op_id Global reduction operator ID
    */
-  void record_reduction_operator(ReductionOpKind op_kind, int64_t global_op_id) const;
+  void record_reduction_operator(ReductionOpKind op_kind, int32_t global_op_id) const;
   /**
    * @brief Finds the global operator ID for a given reduction operator kind.
    *
@@ -179,7 +179,7 @@ class Type {
    *
    * @return Global reduction operator ID
    */
-  [[nodiscard]] int64_t find_reduction_operator(int32_t op_kind) const;
+  [[nodiscard]] int32_t find_reduction_operator(int32_t op_kind) const;
   /**
    * @brief Finds the global operator ID for a given reduction operator kind.
    *
@@ -189,7 +189,7 @@ class Type {
    *
    * @return Global reduction operator ID
    */
-  [[nodiscard]] int64_t find_reduction_operator(ReductionOpKind op_kind) const;
+  [[nodiscard]] int32_t find_reduction_operator(ReductionOpKind op_kind) const;
   /**
    * @brief Equality check between types
    *
